+++ src/main.cpp
@@ -4,11 +4,13 @@
 #include <Geode/loader/Mod.hpp>
 #include <filesystem> 
 
+#include <cmath>
 #include <random>
 std::random_device rd;
 std::mt19937 rng(rd());
 
-static const int profaneSfx[] = { 4, 9, 10, 11, 24, 27 };
+static const int s_profaneSfx[] = { 4, 9, 10, 11, 24, 27 };
+static int s_killerObjectId = -1;
 
 using namespace geode::prelude;
 
@@ -19,8 +21,8 @@ void playSfx(int sfxId, float volume, bool censored) {
     std::string filename = std::to_string(sfxId) + ".ogg";
     if (censored) {
         bool contains = false;
-        for (int i = 0; i < std::size(profaneSfx); i++) {
-            if (profaneSfx[i] == sfxId) {
+        for (int i = 0; i < std::size(s_profaneSfx); i++) {
+            if (s_profaneSfx[i] == sfxId) {
                 contains = true;
                 break;
             }
@@ -89,6 +91,23 @@ class $modify(MyLevelCompleteHook, PlayLayer) {
     }
 };
 
+class $modify(MyGameObjHook, PlayLayer) {
+    struct Fields {
+        int killerObjectId;
+    };
+
+public:
+    void destroyPlayer(PlayerObject* player, GameObject* object) {
+        if (object) {
+            s_killerObjectId = object->m_objectID;
+        }
+        else {
+            s_killerObjectId = -1;
+        }
+        PlayLayer::destroyPlayer(player, object);
+    }
+};
+
 class $modify(MyDeathHook, PlayerObject) {
     struct Fields {
         int scheduledSoundId;
@@ -97,13 +116,14 @@ class $modify(MyDeathHook, PlayerObject) {
     };
     
     public:
+        
         void playDeathEffect() {
             PlayerObject::playDeathEffect();
 
             bool enabled = Mod::get()->getSettingValue<bool>("enable");
             bool disabledInPractice = Mod::get()->getSettingValue<bool>("disableinpractice");
             bool disabledInPlat = Mod::get()->getSettingValue<bool>("disableinplat");
-
+            
             if (!enabled) {
                 return;
             }
@@ -127,7 +147,7 @@ class $modify(MyDeathHook, PlayerObject) {
             bool waveDeath = this->m_isDart;
                         
             m_fields->scheduledVolume = Mod::get()->getSettingValue<int64_t>("volume") / 100.0f;
-            m_fields->scheduledSoundId = chooseDeathSound(deathPercent, currentBest, waveDeath);
+            m_fields->scheduledSoundId = chooseDeathSound(deathPercent, currentBest, waveDeath, s_killerObjectId);
             m_fields->scheduledIsCensored = Mod::get()->getSettingValue<std::string>("censored") == "Censored";
 
             this->scheduleOnce(schedule_selector(MyDeathHook::playMySound), 0.f);
@@ -137,30 +157,39 @@ class $modify(MyDeathHook, PlayerObject) {
             playSfx(m_fields->scheduledSoundId, m_fields->scheduledVolume, m_fields->scheduledIsCensored);
         }
 
-        int chooseDeathSound(int deathPercent, int currentBest, bool waveDeath) {
+        bool isSpikeObject(int objectId) {
+            // i got this list from manually checking Colon's level in-game, not sure if its 100% inclusive but should be mostly accurate
+            static const std::unordered_set<int> spikeIds = { 8, 39, 103, 144, 145, 177, 178, 179, 191, 198, 199, 205, 216, 217, 218, 392, 393, 458, 459, 667, 1703, 1889, 1890, 1891, 1892 };
+            return spikeIds.count(objectId) > 0;
+        }
+
+        int chooseDeathSound(int deathPercent, int currentBest, bool waveDeath, int killerObjectId) {
 
             auto censored = Mod::get()->getSettingValue<std::string>("censored");
-            auto customization = Mod::get()->getSettingValue<std::string>("customize");
-            auto threshold = Mod::get()->getSettingValue<int64_t>("threshold");
+            auto customization = Mod::get()->getSettingValue<std::string>("customize_deaths");
+            auto decentThreshold = Mod::get()->getSettingValue<int64_t>("decent_threshold");
+            auto farThreshold = Mod::get()->getSettingValue<int64_t>("far_threshold");
+            bool alwaysPlayOnBest = Mod::get()->getSettingValue<bool>("alwaysbest");
 
             int rageSize =
-                // small rage: 0% - 25% (regardless of current best)
-                // medium rage: minimum 25%, maximum is the smaller of either threshold or 5% before current best
-                // current best rage: 3% up to and including current best 
-                // large rage: past the best
-                (deathPercent < 25) ? 1 :
-                (deathPercent + 2 <= currentBest && deathPercent < threshold) ? 2 :
-                (deathPercent <= currentBest && deathPercent < threshold) ? 3 :
+                // small rage: 0% - decentThreshold (regardless of current best)
+                // current best rage: 3% up to and including current best
+                // medium rage: minimum decentThreshold, maximum farThreshold
+                // large rage: minimum farThreshold
+                (deathPercent < decentThreshold) ? 1 :
+                (deathPercent >= decentThreshold && std::abs(deathPercent - currentBest) <= 2) ? 3 :
+                (deathPercent >= decentThreshold && deathPercent < farThreshold) ? 2 :
                 4;
 
-            if (customization == "New bests" && deathPercent <= currentBest) {
-                return -1;
+            if (customization == "Decent attempts" && deathPercent <= decentThreshold) {
+                if (!(alwaysPlayOnBest && deathPercent > currentBest)) {
+                    return -1;
+                }
             }
-            if (customization == "Far atts & new bests" && deathPercent <= currentBest && deathPercent <= threshold) {
-                return -1;
-            }
-            if (customization == "Far attempts" && deathPercent <= threshold) {
-                return -1;
+            if (customization == "Far attempts" && deathPercent <= farThreshold) {
+                if (!(alwaysPlayOnBest && deathPercent > currentBest)) {
+                    return -1;
+                }
             }
            
             static const int smallRageSfx[] = { 2, 3, 5, 26 };
@@ -168,7 +197,7 @@ class $modify(MyDeathHook, PlayerObject) {
             static const int largeRageSfx[] = { 1, 4, 6, 12, 13, 15, 17, 20, 21, 22, 23 };
             static const int currentBestRageSfx[] = { 7, 16, 18 };
             static const int spikeDeathRageSfx[] = { 10, 19 };
-            static const int waveDeathRageSfx[] = { 14, 14};
+            static const int waveDeathRageSfx[] = { 14 };
 
             std::vector<int> pool;
 
@@ -193,17 +222,14 @@ class $modify(MyDeathHook, PlayerObject) {
                 case 4: add(largeRageSfx); break;
             }
 
-            if (rageSize == 4) {
-                // was gonna make this only activate when the player actually dies to a spike
-                // but cant figure out how to do that (this is my first geode mod lol
-                // feel free to make a pr if you are more experienced than me
+            if (isSpikeObject(killerObjectId) && rageSize >= 2) {
                 add(spikeDeathRageSfx);
             }
             if (waveDeath && rageSize >= 2) {
                 add(waveDeathRageSfx);
             }
             if (censored == "Disabled") {
-                removeAll(profaneSfx);
+                removeAll(s_profaneSfx);
             }
 
             std::uniform_int_distribution<> dis(0, pool.size() - 1);