+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.2.0
+- Added more customization options ("decent" versus "far" attempts)
+- Spike-related SFX now only play if the player actually died to a spike
+
 # 1.1.0
 - Added an option to disable or censor profane sound effects (on by default)
 - Added an option to disable the mod in Platformer levels